pyproject.toml
[tool.poetry]
name = "hackathon"
version = "0.1.0"
description = "sc-summit-saude-hacking-health-hackathon"
authors = ["bentoluizv <bentoluizv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13.0"
fastapi = {extras = ["standard"], version = "^0.115.4"}
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.6.1"
python-ulid = {extras = ["pydantic"], version = "^3.0.0"}
email-validator = "^2.2.0"
langchain = "^0.3.7"
langchain-cohere = "^0.3.1"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.0"
ruff = "^0.7.3"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
exclude = ["app/database/migrations"]



[tool.ruff.lint]
preview = true
ignore = ['PLR6301']
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
dev = 'fastapi dev hackathon/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=hackathon'
post_test = 'coverage html'
run = 'fastapi run hackathon/app.py'
